Fix packaging error. Release 0.16 (2013-02-18).
This change never made it to master.
I made branch 0.16.x from commit aa08af13370d322ac596b1dfaf56182b3f71b482
(fix tests to work with buildout 2).
This new commit is the diff betweeb that commit and what was released as 0.16 on PyPI.
There is no real change, except the version number, so I don't see what the packaging problem was,
so we don't need to change anything on master.

Anyway, tag 0.16 was not on github, and now it is.